# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Teleconsultation API (with JWT)
  version: "1.1"
  description: API for Teleconsultation PoC including JWT auth (access + refresh).
servers:
  - url: http://localhost:3000
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [PATIENT, DOCTOR, ADMIN]
      responses:
        '201':
          description: Created
  /api/auth/login:
    post:
      summary: Login and receive access + refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: integer }
                      email: { type: string }
                      role: { type: string }
                  accessToken: { type: string }
                  refreshToken: { type: string }
  /api/auth/refresh:
    post:
      summary: Exchange refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
  /api/auth/logout:
    post:
      summary: Revoke refresh token (PoC)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: OK
  /api/profile:
    get:
      summary: Get current user's profile (protected)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  email: { type: string }
                  role: { type: string }
  /api/doctors:
    get:
      summary: List doctors
      responses:
        '200':
          description: OK
  /api/appointments:
    post:
      summary: Create appointment (protected)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patientId: { type: integer }
                doctorId: { type: integer }
                startTime: { type: string, format: date-time }
      responses:
        '201':
          description: Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
externalDocs:
  description: Original project brief (local file)
  url: file:///mnt/data/projet.pdf